pyproject.toml
[project]
name = "sp500-backtest"
version = "0.1.0"
description = "S&P 500 QuantStats 백테스팅 시스템 - Pine Script 지표 재구현 및 조합 최적화"
requires-python = ">=3.12"
dependencies = [
    "yfinance>=0.2.31",
    "quantstats>=0.0.62",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pandas-ta>=0.3.14b1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.82.0",
    "pytest>=7.4.0",
]

[tool.pytest.ini_options]
testpaths = ["sp500_backtest/tests"]
pythonpath = ["."]
